Parser component refactored
Packages reorganized.

CensusParser interface has been renamed FileReader.
CensusParser is now a new class acting as the core of the Parser
component and manages the input reader, the letter writer and the
password generator.

ArgumentsParser now processes the file argument, then creates
CensusParser object and passes the file to it.

The LetterGenertors classes now use Voter as argument instead of name
and email and the passwords are generated preiously by the CensusParser.

Added empty test classes.